+++ apps/modernization-ui/src/apps/patient/add/extended/inputs/phone/PhoneEntryView.spec.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import { PhoneEmailEntry } from 'apps/patient/data/entry';
+import { PhoneEmailEntry } from 'apps/patient/data';
 import { PhoneEntryView } from './PhoneEntryView';
 
 const mockPatientPhoneCodedValues = {
