+++ apps/modernization-ui/src/apps/patient/data/phoneEmail/index.ts
@@ -1,0 +1,5 @@
+export { initial } from './entry';
+export type { PhoneEmailEntry } from './entry';
+
+export { PhoneEmailEntryFields } from './PhoneEmailEntryFields';
+export { asPhoneEmail } from './asPhoneEmail';
